Adding Javascript files via upload
Scripts.js and Marked.min.js needed a few changes to ensure that there's dynamic linking between these files and index.html. If not for this then there would be missing displays of content. Without this, there would also be an issue with the smooth scrolling feature between the sections while navigating the top header.