Revision to index.html
Changed the html to ensure that the hero section (with the background image) is H1 text. Before, it was H2 text. If by any chance this report is released publicly, it's important that the title of the report is an H1 to make sure it is SEO optimized and searchable for people.